Minor changes
Added a default .tif extension if a raster name leaves the extension
unspecified. Also modified the parsing of integer tool arguments so
that it handles floating point string representations.

+++ src/tools/image_analysis/adaptive_filter.rs
@@ -160,22 +160,22 @@ impl WhiteboxTool for AdaptiveFilter {
                 }
             } else if vec[0].to_lowercase() == "-filter" || vec[0].to_lowercase() == "--filter" {
                 if keyval {
-                    filter_size_x = vec[1].to_string().parse::<usize>().unwrap();
+                    filter_size_x = vec[1].to_string().parse::<f32>().unwrap() as usize;
                 } else {
-                    filter_size_x = args[i + 1].to_string().parse::<usize>().unwrap();
+                    filter_size_x = args[i + 1].to_string().parse::<f32>().unwrap() as usize;
                 }
                 filter_size_y = filter_size_x;
             } else if vec[0].to_lowercase() == "-filterx" || vec[0].to_lowercase() == "--filterx" {
                 if keyval {
-                    filter_size_x = vec[1].to_string().parse::<usize>().unwrap();
+                    filter_size_x = vec[1].to_string().parse::<f32>().unwrap() as usize;
                 } else {
-                    filter_size_x = args[i + 1].to_string().parse::<usize>().unwrap();
+                    filter_size_x = args[i + 1].to_string().parse::<f32>().unwrap() as usize;
                 }
             } else if vec[0].to_lowercase() == "-filtery" || vec[0].to_lowercase() == "--filtery" {
                 if keyval {
-                    filter_size_y = vec[1].to_string().parse::<usize>().unwrap();
+                    filter_size_y = vec[1].to_string().parse::<f32>().unwrap() as usize;
                 } else {
-                    filter_size_y = args[i + 1].to_string().parse::<usize>().unwrap();
+                    filter_size_y = args[i + 1].to_string().parse::<f32>().unwrap() as usize;
                 }
             } else if vec[0].to_lowercase() == "-threshold" || vec[0].to_lowercase() == "--threshold" {
                 if keyval {
