+++ src/tools/image_analysis/modified_k_means_clustering.rs
@@ -221,9 +221,9 @@ impl WhiteboxTool for ModifiedKMeansClustering {
                 }
             } else if flag_val == "-start_clusters" {
                 if keyval {
-                    num_classes = vec[1].to_string().parse::<usize>().unwrap();
+                    num_classes = vec[1].to_string().parse::<f32>().unwrap() as usize;
                 } else {
-                    num_classes = args[i + 1].to_string().parse::<usize>().unwrap();
+                    num_classes = args[i + 1].to_string().parse::<f32>().unwrap() as usize;
                 }
             } else if flag_val == "-merger_dist" {
                 if keyval {
@@ -234,9 +234,9 @@ impl WhiteboxTool for ModifiedKMeansClustering {
                 merger_dist *= merger_dist;
             } else if flag_val == "-max_iterations" {
                 if keyval {
-                    max_iterations = vec[1].to_string().parse::<usize>().unwrap();
+                    max_iterations = vec[1].to_string().parse::<f32>().unwrap() as usize;
                 } else {
-                    max_iterations = args[i + 1].to_string().parse::<usize>().unwrap();
+                    max_iterations = args[i + 1].to_string().parse::<f32>().unwrap() as usize;
                 }
             } else if flag_val == "-class_change" {
                 if keyval {
@@ -256,9 +256,9 @@ impl WhiteboxTool for ModifiedKMeansClustering {
             //     }
             // } else if flag_val == "-min_class_size" {
             //     if keyval {
-            //         min_class_size = vec[1].to_string().parse::<usize>().unwrap();
+            //         min_class_size = vec[1].to_string().parse::<f32>().unwrap() as usize;
             //     } else {
-            //         min_class_size = args[i + 1].to_string().parse::<usize>().unwrap();
+            //         min_class_size = args[i + 1].to_string().parse::<f32>().unwrap() as usize;
             //     }
             }
         }
