+++ src/tools/stream_network_analysis/extract_valleys.rs
@@ -192,9 +192,9 @@ impl WhiteboxTool for ExtractValleys {
                 line_thin = true;
             } else if vec[0].to_lowercase() == "-filter" || vec[0].to_lowercase() == "--filter" {
                 if keyval {
-                    filter_size = vec[1].to_string().parse::<usize>().unwrap();
+                    filter_size = vec[1].to_string().parse::<f32>().unwrap() as usize;
                 } else {
-                    filter_size = args[i + 1].to_string().parse::<usize>().unwrap();
+                    filter_size = args[i + 1].to_string().parse::<f32>().unwrap() as usize;
                 }
 
                 //the filter dimensions must be odd numbers such that there is a middle pixel
