+++ src/tools/terrain_analysis/fill_missing_data.rs
@@ -146,9 +146,9 @@ impl WhiteboxTool for FillMissingData {
                 }
             } else if vec[0].to_lowercase() == "-filter" || vec[0].to_lowercase() == "--filter" {
                 if keyval {
-                    filter_size = vec[1].to_string().parse::<usize>().unwrap();
+                    filter_size = vec[1].to_string().parse::<f32>().unwrap() as usize;
                 } else {
-                    filter_size = args[i+1].to_string().parse::<usize>().unwrap();
+                    filter_size = args[i+1].to_string().parse::<f32>().unwrap() as usize;
                 }
             }
         }
