+++ src/tools/terrain_analysis/remove_off_terrain_objects.rs
@@ -154,9 +154,9 @@ impl WhiteboxTool for RemoveOffTerrainObjects {
                 }
             } else if vec[0].to_lowercase() == "-filter" || vec[0].to_lowercase() == "--filter" {
                 if keyval {
-                    filter_size = vec[1].to_string().parse::<usize>().unwrap();
+                    filter_size = vec[1].to_string().parse::<f32>().unwrap() as usize;
                 } else {
-                    filter_size = args[i+1].to_string().parse::<usize>().unwrap();
+                    filter_size = args[i+1].to_string().parse::<f32>().unwrap() as usize;
                 }
             } else if vec[0].to_lowercase() == "-slope" || vec[0].to_lowercase() == "--slope" {
                 if keyval {
